+++ dataset/groupby.cpp
@@ -39,15 +39,24 @@ T GroupBy<T>::copy(const scipp::index group,
     size += slice.end() - slice.begin();
   // This is just the slicing dim, but `slices` may be empty
   const Dim slice_dim = m_data.coords()[dim()].dims().inner();
-  auto out =
-      scipp::dataset::copy(m_data.slice({slice_dim, 0, size}), attrPolicy);
+  auto out = dataset::copy(m_data.slice({slice_dim, 0, size}), attrPolicy);
   scipp::index current = 0;
-  for (const auto &slice : slices) {
+  auto out_slices = slices;
+  for (auto &slice : out_slices) {
     const auto thickness = slice.end() - slice.begin();
-    const Slice out_slice(slice_dim, current, current + thickness);
-    scipp::dataset::copy(m_data.slice(slice), out.slice(out_slice), attrPolicy);
+    slice = Slice(slice.dim(), current, current + thickness);
     current += thickness;
   }
+  const auto copy_slice = [&](const auto &range) {
+    for (scipp::index i = range.begin(); i != range.end(); ++i) {
+      const auto &slice = slices[i];
+      const auto &out_slice = out_slices[i];
+      scipp::dataset::copy(m_data.slice(slice), out.slice(out_slice),
+                           attrPolicy);
+    }
+  };
+  core::parallel::parallel_for(core::parallel::blocked_range(0, slices.size()),
+                               copy_slice);
   return out;
 }