+++ variable/include/scipp/variable/misc_operations.h
@@ -11,12 +11,6 @@ namespace scipp::variable {
 
 SCIPP_VARIABLE_EXPORT Variable astype(const Variable &var, const DType type);
 
-SCIPP_VARIABLE_EXPORT std::vector<Variable>
-split(const Variable &var, const Dim dim,
-      const std::vector<scipp::index> &indices);
-SCIPP_VARIABLE_EXPORT Variable filter(const Variable &var,
-                                      const Variable &filter);
-
 SCIPP_VARIABLE_EXPORT Variable masked_to_zero(const Variable &var,
                                               const Variable &mask);